Provide full indexed schema during walk
