+++ crates/constraints/src/utils.rs
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+//! Shared helpers for constraint implementations.
+
+use std::fmt::Display;
+
+/// Join errors into a bulleted, newline-separated list.
+pub fn join_errors<E: Display>(errors: &[E]) -> String {
+    errors
+        .iter()
+        .map(|e| format!("  - {e}"))
+        .collect::<Vec<_>>()
+        .join("\n")
+}
